Add flag for bounds checking
