+++ src-unchecked/Data/Primitive/ByteArray/Operations.hs
@@ -1,0 +1,6 @@
+module Data.Primitive.ByteArray.Operations (
+  readByteArray, writeByteArray, indexByteArray,
+  readCharArray, writeCharArray, indexCharArray,
+) where
+
+import Data.Primitive.ByteArray.Internal
